@@ -2,8 +2,8 @@ package com.starcapital.collectapp.adapters;
 
 import android.content.Context;
 import android.content.Intent;
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -14,7 +14,6 @@ import com.starcapital.collectapp.activities.AccountDetailsActivity;
 import com.starcapital.collectapp.models.Account;
 
 import java.util.ArrayList;
-import java.util.zip.Inflater;
 
 public class AccountsRecyclerAdapter extends RecyclerView.Adapter<AccountsRecyclerAdapter.ViewHolder> {
 
